LayoutTests/http/tests/webgpu/webgpu/webgpu/shader/execution/statement/discard-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: map async was not successful
CONSOLE MESSAGE: Unhandled Promise Rejection: OperationError: map async was not successful

Harness Error (FAIL), message = Unhandled rejection: map async was not successful

FAIL :all:useStorageBuffers=false assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
    assert@http://127.0.0.1:8000/webgpu/common/util/util.js:37:20
    release@http://127.0.0.1:8000/webgpu/webgpu/util/device_pool.js:106:13
 Reached unreachable code
FAIL :all:useStorageBuffers=true assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :three_quarters:useStorageBuffers=false assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
    assert@http://127.0.0.1:8000/webgpu/common/util/util.js:37:20
    release@http://127.0.0.1:8000/webgpu/webgpu/util/device_pool.js:106:13
 Reached unreachable code
FAIL :three_quarters:useStorageBuffers=true assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :function_call:useStorageBuffers=false assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
    assert@http://127.0.0.1:8000/webgpu/common/util/util.js:37:20
    release@http://127.0.0.1:8000/webgpu/webgpu/util/device_pool.js:106:13
 Reached unreachable code
FAIL :function_call:useStorageBuffers=true assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :loop:useStorageBuffers=false assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
    assert@http://127.0.0.1:8000/webgpu/common/util/util.js:37:20
    release@http://127.0.0.1:8000/webgpu/webgpu/util/device_pool.js:106:13
 Reached unreachable code
FAIL :loop:useStorageBuffers=true assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :continuing:useStorageBuffers=false assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
    assert@http://127.0.0.1:8000/webgpu/common/util/util.js:37:20
    release@http://127.0.0.1:8000/webgpu/webgpu/util/device_pool.js:106:13
 Reached unreachable code
FAIL :continuing:useStorageBuffers=true assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :uniform_read_loop:useStorageBuffers=false assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
    assert@http://127.0.0.1:8000/webgpu/common/util/util.js:37:20
    release@http://127.0.0.1:8000/webgpu/webgpu/util/device_pool.js:106:13
 Reached unreachable code
FAIL :uniform_read_loop:useStorageBuffers=true assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
FAIL :derivatives:useStorageBuffers=false assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
    assert@http://127.0.0.1:8000/webgpu/common/util/util.js:37:20
    release@http://127.0.0.1:8000/webgpu/webgpu/util/device_pool.js:106:13
 Reached unreachable code
FAIL :derivatives:useStorageBuffers=true assert_unreached:
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: OperationError: map async was not successful
  - EXCEPTION: Error: Device was unexpectedly lost. Reason: unknown, Message:
      at (elided: only 2 shown)
 Reached unreachable code
